add NavigationContainer
